requirements.txt
requests == 2.31.0
python-dotenv == 1.0.0
pandas == 2.2.3
pyarrow == 18.1.0
google-cloud-bigquery == 3.25.0
db-dtypes == 1.2.0
google-auth == 2.28.0
dbt-bigquery == 1.7.0
loguru == 0.7.2
tenacity == 8.2.3
pydantic == 2.10.0
pytest == 8.3.5
